Fix map loading

+++ src/map/house/house.h
@@ -158,10 +158,10 @@ class House
 			return rentWarnings;
 		}
 
-		void setTownId(uint32_t newTownId) {
+		void setTownId(uint8_t newTownId) {
 			this->townId = newTownId;
 		}
-		uint32_t getTownId() const {
+		uint8_t getTownId() const {
 			return townId;
 		}
 
@@ -219,7 +219,7 @@ class House
 		uint32_t ownerAccountId = 0;
 		uint32_t rentWarnings = 0;
 		uint32_t rent = 0;
-		uint32_t townId = 0;
+		uint8_t townId = 0;
 
 		Position posEntry = {};
